pyproject.toml
[project]
name = "moccasin_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "moccasin>=0.4.0",
    "titanoboa==0.2.6",
    "vyper==0.4.1",
]
